+++ mne/viz/utils.py
@@ -19,7 +19,7 @@ import tempfile
 import numpy as np
 
 from ..io import show_fiff
-from ..utils import verbose, get_config, set_config
+from ..utils import verbose, set_config
 
 
 COLORS = ['b', 'g', 'r', 'c', 'm', 'y', 'k', '#473C8B', '#458B74',
@@ -475,85 +475,6 @@ def figure_nobar(*args, **kwargs):
     return fig
 
 
-def _prepare_mne_browse_raw(params, title, bgcolor, color, bad_color, inds,
-                            n_channels):
-    """Helper for setting up the mne_browse_raw window."""
-    import matplotlib.pyplot as plt
-    import matplotlib as mpl
-    size = get_config('MNE_BROWSE_RAW_SIZE')
-    if size is not None:
-        size = size.split(',')
-        size = tuple([float(s) for s in size])
-
-    fig = figure_nobar(facecolor=bgcolor, figsize=size)
-    fig.canvas.set_window_title('mne_browse_raw')
-    ax = plt.subplot2grid((10, 10), (0, 1), colspan=8, rowspan=9)
-    ax.set_title(title, fontsize=12)
-    ax_hscroll = plt.subplot2grid((10, 10), (9, 1), colspan=8)
-    ax_hscroll.get_yaxis().set_visible(False)
-    ax_hscroll.set_xlabel('Time (s)')
-    ax_vscroll = plt.subplot2grid((10, 10), (0, 9), rowspan=9)
-    ax_vscroll.set_axis_off()
-    ax_help_button = plt.subplot2grid((10, 10), (0, 0), colspan=1)
-    help_button = mpl.widgets.Button(ax_help_button, 'Help')
-    help_button.on_clicked(partial(_onclick_help, params=params))
-    # store these so they can be fixed on resize
-    params['fig'] = fig
-    params['ax'] = ax
-    params['ax_hscroll'] = ax_hscroll
-    params['ax_vscroll'] = ax_vscroll
-    params['ax_help_button'] = ax_help_button
-    params['help_button'] = help_button
-
-    # populate vertical and horizontal scrollbars
-    info = params['info']
-    for ci in range(len(info['ch_names'])):
-        this_color = (bad_color if info['ch_names'][inds[ci]] in info['bads']
-                      else color)
-        if isinstance(this_color, dict):
-            this_color = this_color[params['types'][inds[ci]]]
-        ax_vscroll.add_patch(mpl.patches.Rectangle((0, ci), 1, 1,
-                                                   facecolor=this_color,
-                                                   edgecolor=this_color))
-    vsel_patch = mpl.patches.Rectangle((0, 0), 1, n_channels, alpha=0.5,
-                                       facecolor='w', edgecolor='w')
-    ax_vscroll.add_patch(vsel_patch)
-    params['vsel_patch'] = vsel_patch
-    hsel_patch = mpl.patches.Rectangle((params['t_start'], 0),
-                                       params['duration'], 1, edgecolor='k',
-                                       facecolor=(0.75, 0.75, 0.75),
-                                       alpha=0.25, linewidth=1, clip_on=False)
-    ax_hscroll.add_patch(hsel_patch)
-    params['hsel_patch'] = hsel_patch
-    ax_hscroll.set_xlim(0, params['n_times'] / float(info['sfreq']))
-    n_ch = len(info['ch_names'])
-    ax_vscroll.set_ylim(n_ch, 0)
-    ax_vscroll.set_title('Ch.')
-
-    # make shells for plotting traces
-    ylim = [n_channels * 2 + 1, 0]
-    offset = ylim[0] / n_channels
-    offsets = np.arange(n_channels) * offset + (offset / 2.)
-    ax.set_yticks(offsets)
-    ax.set_ylim(ylim)
-    ax.set_xlim(params['t_start'], params['t_start'] + params['duration'],
-                False)
-
-    params['offsets'] = offsets
-    params['lines'] = [ax.plot([np.nan], antialiased=False, linewidth=0.5)[0]
-                       for _ in range(n_ch)]
-    ax.set_yticklabels(['X' * max([len(ch) for ch in info['ch_names']])])
-    vertline_color = (0., 0.75, 0.)
-    params['ax_vertline'] = ax.plot([0, 0], ylim, color=vertline_color,
-                                    zorder=-1)[0]
-    params['ax_vertline'].ch_name = ''
-    params['vertline_t'] = ax_hscroll.text(0, 1, '', color=vertline_color,
-                                           va='bottom', ha='right')
-    params['ax_hscroll_vertline'] = ax_hscroll.plot([0, 0], [0, 1],
-                                                    color=vertline_color,
-                                                    zorder=1)[0]
-
-
 def _helper_raw_resize(event, params):
     """Helper for resizing"""
     size = ','.join([str(s) for s in params['fig'].get_size_inches()])
@@ -729,87 +650,6 @@ def _select_bads(event, params, bads):
         params['ax_hscroll_vertline'].set_data(x, np.array([0., 1.]))
         params['vertline_t'].set_text('%0.3f' % x[0])
     return bads
-
-
-def _plot_raw_traces(params, inds, color, bad_color, event_lines=None,
-                     event_color=None):
-    """Helper for plotting raw"""
-    lines = params['lines']
-    info = params['info']
-    n_channels = params['n_channels']
-    params['bad_color'] = bad_color
-    # do the plotting
-    tick_list = []
-    for ii in range(n_channels):
-        ch_ind = ii + params['ch_start']
-        # let's be generous here and allow users to pass
-        # n_channels per view >= the number of traces available
-        if ii >= len(lines):
-            break
-        elif ch_ind < len(info['ch_names']):
-            # scale to fit
-            ch_name = info['ch_names'][inds[ch_ind]]
-            tick_list += [ch_name]
-            offset = params['offsets'][ii]
-
-            # do NOT operate in-place lest this get screwed up
-            this_data = params['data'][inds[ch_ind]] * params['scale_factor']
-            this_color = bad_color if ch_name in info['bads'] else color
-            this_z = -1 if ch_name in info['bads'] else 0
-            if isinstance(this_color, dict):
-                this_color = this_color[params['types'][inds[ch_ind]]]
-
-            # subtraction here gets corect orientation for flipped ylim
-            lines[ii].set_ydata(offset - this_data)
-            lines[ii].set_xdata(params['times'])
-            lines[ii].set_color(this_color)
-            lines[ii].set_zorder(this_z)
-            vars(lines[ii])['ch_name'] = ch_name
-            vars(lines[ii])['def_color'] = color[params['types'][inds[ch_ind]]]
-        else:
-            # "remove" lines
-            lines[ii].set_xdata([])
-            lines[ii].set_ydata([])
-    # deal with event lines
-    if params['event_times'] is not None:
-        # find events in the time window
-        event_times = params['event_times']
-        mask = np.logical_and(event_times >= params['times'][0],
-                              event_times <= params['times'][-1])
-        event_times = event_times[mask]
-        event_nums = params['event_nums'][mask]
-        # plot them with appropriate colors
-        # go through the list backward so we end with -1, the catchall
-        used = np.zeros(len(event_times), bool)
-        ylim = params['ax'].get_ylim()
-        for ev_num, line in zip(sorted(event_color.keys())[::-1],
-                                event_lines[::-1]):
-            mask = (event_nums == ev_num) if ev_num >= 0 else ~used
-            assert not np.any(used[mask])
-            used[mask] = True
-            t = event_times[mask]
-            if len(t) > 0:
-                xs = list()
-                ys = list()
-                for tt in t:
-                    xs += [tt, tt, np.nan]
-                    ys += [0, ylim[0], np.nan]
-                line.set_xdata(xs)
-                line.set_ydata(ys)
-            else:
-                line.set_xdata([])
-                line.set_ydata([])
-    # finalize plot
-    params['ax'].set_xlim(params['times'][0],
-                          params['times'][0] + params['duration'], False)
-    params['ax'].set_yticklabels(tick_list)
-    params['vsel_patch'].set_y(params['ch_start'])
-    params['fig'].canvas.draw()
-    # XXX This is a hack to make sure this figure gets drawn last
-    # so that when matplotlib goes to calculate bounds we don't get a
-    # CGContextRef error on the MacOSX backend :(
-    if params['fig_proj'] is not None:
-        params['fig_proj'].canvas.draw()
 
 
 def _onclick_help(event, params):
